refactor(repo): resolve residual style-gate findings — circuit-breaker split, redundant overrides, module-owner dispositions
