+++ apps/api-v1/src/middleware-resilience.ts
@@ -1,6 +1,6 @@
 import { Temporal } from '@js-temporal/polyfill';
 import { ResilienceDto } from '@shared/queuebox/DequeueResourceEntryController.ts';
-import { CircuitBreakerPolicy } from '@shared/resilience/Resilience.ts';
+import { CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 
 const duration = Temporal.Duration.from({ seconds: 10 });
 const MAX_FAIRNESS_SELECTIONS_PER_WINDOW = 10;
