+++ packages/shared-server/rallar-system/services/AppCrdtInboxService.ts
@@ -54,11 +54,7 @@ export class AppCrdtInboxService extends AppInboxService {
   private readonly wakeQueueEngine: (() => void) | undefined;
   private readonly resolveCurrentSession: ResolveCurrentCrdtMutationSession | undefined;
 
-  public override readonly inbox: InboxQueueReader;
-  public override readonly resourceInbox: ResourceInboxRepository;
-  public override readonly resourceInboxResults: ResourceInboxResultsRepository;
   public readonly mutationService: CrdtMutationService;
-  public override readonly serviceId: string;
 
   constructor(
     inbox: InboxQueueReader,
@@ -87,11 +83,7 @@ export class AppCrdtInboxService extends AppInboxService {
       options,
       effects.wakeQueueEngine,
     );
-    this.inbox = inbox;
-    this.resourceInbox = resourceInbox;
-    this.resourceInboxResults = resourceInboxResults;
     this.mutationService = mutationService;
-    this.serviceId = serviceId;
     this.audit = effects.audit;
     this.wakeQueueEngine = effects.wakeQueueEngine;
     this.resolveCurrentSession = effects.resolveCurrentSession;
