+++ packages/shared-web/browser/resilience-config.ts
@@ -1,5 +1,5 @@
 import { Temporal } from '@js-temporal/polyfill';
-import { CircuitBreakerPolicy } from '@shared/resilience/Resilience.ts';
+import { CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 import { ResilienceDto } from '@shared/queuebox/DequeueResourceEntryController.ts';
 
 const duration = Temporal.Duration.from({ seconds: 10 });
