+++ packages/tests/shared-server/client-state/app-client-inbox-mutation-test-harness.ts
@@ -12,7 +12,7 @@ import {
   type ResourceEntry,
   toKeyAsString,
 } from '@shared/queuebox/ResourceEntry.ts';
-import { CircuitBreakerPolicy } from '@shared/resilience/Resilience.ts';
+import { CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 import { Either } from '@shared/resilience/Either.ts';
 import { InboxQueueReader } from '@shared/services/InboxQueueReader.ts';
 import {
