+++ packages/tests/shared-server/fixtures/postgres-app-inbox-worker-runtime.ts
@@ -1,6 +1,6 @@
 import { Temporal } from '@js-temporal/polyfill';
 import type { StateScope } from '@shared/api/state-types.ts';
-import { CircuitBreakerPolicy } from '@shared/resilience/Resilience.ts';
+import { CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 import type { Either } from '@shared/resilience/Either.ts';
 import { ResilienceDto } from '@shared/queuebox/DequeueResourceEntryController.ts';
 import type { ResourceInboxAttemptReleaseTelemetry } from '@shared/queuebox/ResourceInboxAttemptTelemetry.ts';
