+++ packages/tests/shared-server/fixtures/postgres-topology-app-outbox-worker.ts
@@ -2,7 +2,7 @@ import { Temporal } from '@js-temporal/polyfill';
 import postgres from 'postgres';
 
 import type { GroupSnapshot } from '@shared/api/group-types.ts';
-import { CircuitBreakerPolicy } from '@shared/resilience/Resilience.ts';
+import { CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 import { ResilienceDto } from '@shared/queuebox/DequeueResourceEntryController.ts';
 import { EntityStatus, type Key } from '@shared/queuebox/ResourceEntry.ts';
 import { OutboxQueueReader } from '@shared/services/OutboxQueueReader.ts';
