+++ packages/tests/shared/resilience.test.ts
@@ -1,8 +1,7 @@
 import { Temporal } from '@js-temporal/polyfill';
 import { afterEach, describe, expect, it, vi } from 'vitest';
+import { CircuitBreaker, CircuitBreakerPolicy } from '@shared/resilience/circuit-breaker.ts';
 import {
-    CircuitBreaker,
-    CircuitBreakerPolicy,
     RateAdjuster,
     RateLimiter,
     SlidingWindowCounter,
